+++ aider/__init__.py
@@ -1,6 +1,6 @@
 from packaging import version
 
-__version__ = "0.66.1.dev"
+__version__ = "0.69.2.dev"
 safe_version = __version__
 
 try: